+++ packages/@headlessui-react/src/components/tooltip/tooltip.tsx
@@ -23,6 +23,7 @@ import {
   FloatingProvider,
   useFloatingPanel,
   useFloatingReference,
+  useResolvedAnchor,
   type AnchorProps,
 } from '../../internal/floating'
 import { State, useOpenClosed } from '../../internal/open-closed'
@@ -422,15 +423,7 @@ function PanelFn<TTag extends ElementType = typeof DEFAULT_PANEL_TAG>(
   props: TooltipPanelProps<TTag>,
   ref: Ref<HTMLElement>
 ) {
-  let {
-    anchor = {
-      to: 'top',
-      padding: 8,
-      gap: 8,
-      offset: -4,
-    } as AnchorProps,
-    ...theirProps
-  } = props
+  let { anchor: rawAnchor, ...theirProps } = props
   let data = useData('TooltipPanel')
 
   let usesOpenClosedState = useOpenClosed()
@@ -443,6 +436,7 @@ function PanelFn<TTag extends ElementType = typeof DEFAULT_PANEL_TAG>(
   })()
 
   let internalPanelRef = useRef<HTMLElement | null>(null)
+  let anchor = useResolvedAnchor(rawAnchor ?? { to: 'top', padding: 8, gap: 8, offset: -4 })
   let [floatingRef, style] = useFloatingPanel(visible ? anchor : undefined)
   let panelRef = useSyncRefs(internalPanelRef, ref, floatingRef)